Add pagination to get credit rates, update credit wallet conversion model

+++ bigfastapi/schemas/credit_wallet_conversion_schemas.py
@@ -3,7 +3,7 @@ import pydantic as _pydantic
 
 class CreditWalletConversion(_pydantic.BaseModel):
     credit_wallet_type: str
-    rate: int
+    rate: float
     currency_code: str
 
     class Config:
